Prepare for initial release
- Separate cli and bluehawk lib sources
- Update package.json to include useful information
