+++ src/bluehawk/BluehawkError.ts
