+++ src/bluehawk/Document.ts
