+++ src/bluehawk/commands/Command.ts
