+++ src/bluehawk/commands/RemoveCommand.ts
