+++ src/bluehawk/commands/removeCommand.test.ts
