+++ src/bluehawk/commands/replaceCommand.test.ts
