+++ src/bluehawk/commands/stateCommand.test.ts
