+++ src/bluehawk/parser/CommandNode.ts
