+++ src/bluehawk/parser/lexer/makePayloadPattern.ts
