+++ src/bluehawk/parser/lexer/makeRootMode.ts
