+++ src/bluehawk/parser/lexer/tokenCategoryFilter.ts
