+++ src/bluehawk/parser/visitor/innerOffsetToOuterLocation.ts
