+++ src/bluehawk/processor/Processor.ts
