+++ src/bluehawk/processor/makeAttributesConformToJsonSchemaRule.ts
