+++ src/bluehawk/processor/validator.ts
