+++ src/cli/index.ts
@@ -1,0 +1,7 @@
+#!/usr/bin/env node
+
+import { run } from "./cli";
+
+run().catch((err) => {
+  console.error(err);
+});
